Task #192 Create REST api to delete user from the database

+++ src/Model/user.py
@@ -51,8 +51,11 @@ class UserMode:
     @classmethod
     def DeleteUser(cls, username):
         user = sql.User.query.filter_by(Username=username).first()
+        if not user:
+            return False
         sql.db.session.delete(user)
         sql.db.session.commit()
+        return True
 
     @classmethod
     def UpdateGrade(cls, username, grade):
@@ -61,6 +64,7 @@ class UserMode:
             return False
         user.Grade = grade
         sql.db.session.commit()
+        
         return True
